requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==5.0.3
django-cors-headers==4.3.1
django-filter==24.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
inflection==0.5.1
kombu==5.3.5
packaging==24.0
pillow==10.2.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-slugify==8.0.4
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.10.0
tzdata==2024.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
channels==4.0.0
